refactor: rename and move icp enums
