+++ packages/nns/src/enums/icp.enums.ts
