+++ packages/nns/src/icp.ts
@@ -1,6 +1,6 @@
 import { ICPTs } from "../proto/ledger_pb";
 import { E8S_PER_ICP } from "./constants/constants";
-import { FromICPStringError } from "./types/icp";
+import { FromICPStringError } from "./enums/icp.enums";
 
 export class ICP {
   private constructor(private e8s: bigint) {}
