+++ packages/nns/src/index.ts
@@ -1,6 +1,7 @@
 export type { DeployedSns } from "../candid/sns_wasm";
 export { AccountIdentifier, SubAccount } from "./account_identifier";
 export * from "./enums/governance.enums";
+export * from "./enums/icp.enums";
 export * from "./errors/governance.errors";
 export * from "./errors/ledger.errors";
 export { GenesisTokenCanister } from "./genesis_token.canister";
@@ -11,7 +12,6 @@ export { SnsWasmCanister } from "./sns_wasm.canister";
 export * from "./types/common";
 export * from "./types/governance.options";
 export * from "./types/governance_converters";
-export * from "./types/icp";
 export * from "./types/ledger.options";
 export type { SnsWasmCanisterOptions } from "./types/sns_wasm.options";
 export * from "./utils/accounts.utils";
